requirements.txt
attrs==21.2.0
awsebcli==3.20.1
bcrypt==3.2.0
blessed==1.18.1
boto3==1.18.9
botocore==1.21.9
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.3
cryptography==3.4.7
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
Flask==2.0.1
Flask-WTF==0.15.1
future==0.16.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
paramiko==2.7.2
pathspec==0.5.9
protobuf==3.17.3
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
semantic-version==2.8.5
six==1.14.0
termcolor==1.1.0
texttable==1.6.4
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==2.0.1
WTForms==2.3.3
